# MapGenerator optimization