--- a/core/src/com/unciv/logic/map/mapgenerator/MapGenerator.kt
+++ b/core/src/com/unciv/logic/map/mapgenerator/MapGenerator.kt
@@ -7,12 +7,18 @@ import com.unciv.logic.map.*
 import com.unciv.models.Counter
 import com.unciv.models.ruleset.Ruleset
 import com.unciv.models.ruleset.tile.ResourceType
+import com.unciv.models.ruleset.tile.Terrain
 import com.unciv.models.ruleset.tile.TerrainType
 import kotlin.math.*
 import kotlin.random.Random
 
 
 class MapGenerator(val ruleset: Ruleset) {
+    companion object {
+        const val consoleOutput = true
+        private const val consoleTimings = true
+    }
+    
     private var randomness = MapGenerationRandomness()
 
     fun generateMap(mapParameters: MapParameters): TileMap {
@@ -40,23 +46,54 @@ class MapGenerator(val ruleset: Ruleset) {
             return map
         }
 
-        MapLandmassGenerator(ruleset, randomness).generateLand(map)
-        raiseMountainsAndHills(map)
-        applyHumidityAndTemperature(map)
-        spawnLakesAndCoasts(map)
-        spawnVegetation(map)
-        spawnRareFeatures(map)
-        spawnIce(map)
-        NaturalWonderGenerator(ruleset, randomness).spawnNaturalWonders(map)
-        RiverGenerator(randomness).spawnRivers(map)
-        spreadResources(map)
-        spreadAncientRuins(map)
+        if (consoleTimings) println("\nMapGenerator run with parameters $mapParameters")
+        runAndMeasure("MapLandmassGenerator") {
+            MapLandmassGenerator(ruleset, randomness).generateLand(map)
+        }
+        runAndMeasure("raiseMountainsAndHills") {
+            raiseMountainsAndHills(map)
+        }
+        runAndMeasure("applyHumidityAndTemperature") {
+            applyHumidityAndTemperature(map)
+        }
+        runAndMeasure("spawnLakesAndCoasts") {
+            spawnLakesAndCoasts(map)
+        }
+        runAndMeasure("spawnVegetation") {
+            spawnVegetation(map)
+        }
+        runAndMeasure("spawnRareFeatures") {
+            spawnRareFeatures(map)
+        }
+        runAndMeasure("spawnIce") {
+            spawnIce(map)
+        }
+        runAndMeasure("NaturalWonderGenerator") {
+            NaturalWonderGenerator(ruleset, randomness).spawnNaturalWonders(map)
+        }
+        runAndMeasure("RiverGenerator") {
+            RiverGenerator(randomness).spawnRivers(map)
+        }
+        runAndMeasure("spreadResources") {
+            spreadResources(map)
+        }
+        runAndMeasure("spreadAncientRuins") {
+            spreadAncientRuins(map)
+        }
         return map
     }
 
-    private fun seedRNG(seed: Long, verbose: Boolean = false) {
+    private fun runAndMeasure(text: String, action: ()->Unit) {
+        if (!consoleTimings) return action()
+        val startNanos = System.nanoTime()
+        action()
+        val delta = System.nanoTime() - startNanos
+        println("MapGenerator.$text took ${delta/1000000L}.${(delta/10000L).rem(100)}ms")
+    }
+    //todo: Why is this unused?
+    private fun seedRNG(seed: Long) {
         randomness.RNG = Random(seed)
-        if (verbose) println("RNG seeded with $seed")
+        if (consoleOutput) println("RNG seeded with $seed")
     }
 
     private fun spawnLakesAndCoasts(map: TileMap) {
@@ -109,23 +146,23 @@ class MapGenerator(val ruleset: Ruleset) {
             return
         val suitableTiles = map.values.filter { it.isLand && !it.isImpassible() }
         val locations = randomness.chooseSpreadOutLocations(suitableTiles.size / 50,
-                suitableTiles, 10)
+                suitableTiles, map.mapParameters.mapSize.radius)
         for (tile in locations)
             tile.improvement = Constants.ancientRuins
     }
 
     private fun spreadResources(tileMap: TileMap) {
-        val distance = tileMap.mapParameters.mapSize.radius
+        val mapRadius = tileMap.mapParameters.mapSize.radius
         for (tile in tileMap.values)
             tile.resource = null
 
-        spreadStrategicResources(tileMap, distance)
-        spreadResources(tileMap, distance, ResourceType.Luxury)
-        spreadResources(tileMap, distance, ResourceType.Bonus)
+        spreadStrategicResources(tileMap, mapRadius)
+        spreadResources(tileMap, mapRadius, ResourceType.Luxury)
+        spreadResources(tileMap, mapRadius, ResourceType.Bonus)
     }
 
     // Here, we need each specific resource to be spread over the map - it matters less if specific resources are near each other
-    private fun spreadStrategicResources(tileMap: TileMap, distance: Int) {
+    private fun spreadStrategicResources(tileMap: TileMap, mapRadius: Int) {
         val strategicResources = ruleset.tileResources.values.filter { it.resourceType == ResourceType.Strategic }
         // passable land tiles (no mountains, no wonders) without resources yet
         val candidateTiles = tileMap.values.filter { it.resource == null && !it.isImpassible() }
@@ -138,7 +175,7 @@ class MapGenerator(val ruleset: Ruleset) {
                     .filter { it.resource == null
                             && resource.terrainsCanBeFoundOn.contains(it.getLastTerrain().name) }
 
-            val locations = randomness.chooseSpreadOutLocations(resourcesPerType, suitableTiles, distance)
+            val locations = randomness.chooseSpreadOutLocations(resourcesPerType, suitableTiles, mapRadius)
 
             for (location in locations) location.resource = resource.name
         }
@@ -148,7 +185,7 @@ class MapGenerator(val ruleset: Ruleset) {
      * Spreads resources of type [resourceType] picking locations at [distance] from each other.
      * [MapParameters.resourceRichness] used to control how many resources to spawn.
      */
-    private fun spreadResources(tileMap: TileMap, distance: Int, resourceType: ResourceType) {
+    private fun spreadResources(tileMap: TileMap, mapRadius: Int, resourceType: ResourceType) {
         val resourcesOfType = ruleset.tileResources.values.filter { it.resourceType == resourceType }
 
         val suitableTiles = tileMap.values
@@ -156,7 +193,7 @@ class MapGenerator(val ruleset: Ruleset) {
                 .filter { it.resource == null && resourcesOfType.any { r -> r.terrainsCanBeFoundOn.contains(it.getLastTerrain().name) } }
         val numberOfResources = tileMap.values.count { it.isLand && !it.isImpassible() } *
                 tileMap.mapParameters.resourceRichness
-        val locations = randomness.chooseSpreadOutLocations(numberOfResources.toInt(), suitableTiles, distance)
+        val locations = randomness.chooseSpreadOutLocations(numberOfResources.toInt(), suitableTiles, mapRadius)
 
         val resourceToNumber = Counter<String>()
 
@@ -185,14 +222,14 @@ class MapGenerator(val ruleset: Ruleset) {
             return
         }
 
-        if (mountain != null)
-            println("Mountainlike generation for " + mountain)
-        if (hill != null)
-            println("Hill-like generation for " + hill)
+        if (consoleOutput && mountain != null)
+            println("Mountain-like generation for $mountain")
+        if (consoleOutput && hill != null)
+            println("Hill-like generation for $hill")
 
         val elevationSeed = randomness.RNG.nextInt().toDouble()
         tileMap.setTransients(ruleset)
-        for (tile in tileMap.values.filter { !it.isWater }) {
+        for (tile in tileMap.values.asSequence().filter { !it.isWater }) {
             var elevation = randomness.getPerlinNoise(tile, elevationSeed, scale = 2.0)
                     elevation = abs(elevation).pow(1.0 - tileMap.mapParameters.elevationExponent.toDouble()) * elevation.sign
 
@@ -262,7 +299,7 @@ class MapGenerator(val ruleset: Ruleset) {
         for (i in 1..5) {
             var totalHills = tileMap.values.count { it.terrainFeatures.contains(hill) }
 
-            for (tile in tileMap.values.filter { !it.isWater && (mountain == null || it.baseTerrain != mountain) }) {
+            for (tile in tileMap.values.asSequence().filter { !it.isWater && (mountain == null || it.baseTerrain != mountain) }) {
                 val adjacentMountains = if (mountain == null) 0 else
                     tile.neighbors.count { it.baseTerrain == mountain }
                 val adjacentHills =
@@ -278,7 +315,7 @@ class MapGenerator(val ruleset: Ruleset) {
 
             }
 
-            for (tile in tileMap.values.filter { !it.isWater && (mountain == null || it.baseTerrain != mountain) }) {
+            for (tile in tileMap.values.asSequence().filter { !it.isWater && (mountain == null || it.baseTerrain != mountain) }) {
                 if (tile.terrainFeatures.remove(Constants.rising) && (totalHills <= targetHills || i == 1) ) {
                     if (!tile.terrainFeatures.contains(hill)) {
                         tile.terrainFeatures.add(hill)
@@ -286,10 +323,8 @@ class MapGenerator(val ruleset: Ruleset) {
                     }
                 }
                 if (tile.terrainFeatures.remove(Constants.lowering) && (totalHills >= targetHills * 0.9f || i == 1)) {
-                    if (tile.terrainFeatures.contains(hill)) {
-                        tile.terrainFeatures.remove(hill)
+                    if (tile.terrainFeatures.remove(hill))
                         totalHills--
-                    }
                 }
             }
         }
@@ -306,9 +341,29 @@ class MapGenerator(val ruleset: Ruleset) {
         tileMap.setTransients(ruleset)
 
         val scale = tileMap.mapParameters.tilesPerBiomeArea.toDouble()
+        val temperatureExtremeness = tileMap.mapParameters.temperatureExtremeness
+        
+        class TerrainOccursRange(
+            val terrain: Terrain,
+            val tempFrom: Float, val tempTo: Float,
+            val humidFrom: Float, val humidTo: Float
+        )
+        val limitsMap: List<TerrainOccursRange> =
+            // List is OK here as it's only sequentially scanned
+            ruleset.terrains.values.flatMap { terrain ->
+                terrain.uniqueObjects.filter {
+                    it.placeholderText == "Occurs at temperature between [] and [] and humidity between [] and []"
+                }.map { unique ->
+                    TerrainOccursRange(terrain,
+                        unique.params[0].toFloat(), unique.params[1].toFloat(),
+                        unique.params[2].toFloat(), unique.params[3].toFloat())
+                }
+            }
+        val noTerrainUniques = limitsMap.isEmpty()
+        val elevationTerrains = arrayOf(Constants.mountain, Constants.hill)
 
-        for (tile in tileMap.values) {
-            if (tile.isWater || tile.baseTerrain in arrayOf(Constants.mountain, Constants.hill))
+        for (tile in tileMap.values.asSequence()) {
+            if (tile.isWater || tile.baseTerrain in elevationTerrains)
                 continue
 
             val humidity = (randomness.getPerlinNoise(tile, humiditySeed, scale = scale, nOctaves = 1) + 1.0) / 2.0
@@ -316,36 +371,32 @@ class MapGenerator(val ruleset: Ruleset) {
             val randomTemperature = randomness.getPerlinNoise(tile, temperatureSeed, scale = scale, nOctaves = 1)
             val latitudeTemperature = 1.0 - 2.0 * abs(tile.latitude) / tileMap.maxLatitude
             var temperature = (5.0 * latitudeTemperature + randomTemperature) / 6.0
-            temperature = abs(temperature).pow(1.0 - tileMap.mapParameters.temperatureExtremeness) * temperature.sign
+            temperature = abs(temperature).pow(1.0 - temperatureExtremeness) * temperature.sign
 
             // Old, static map generation rules - necessary for existing base ruleset mods to continue to function
-            if (ruleset.terrains.values.asSequence().flatMap { it.uniqueObjects }
-                            .none { it.placeholderText == "Occurs at temperature between [] and [] and humidity between [] and []" }) {
+            if (noTerrainUniques) {
                 tile.baseTerrain = when {
                     temperature < -0.4 -> if (humidity < 0.5) Constants.snow   else Constants.tundra
                     temperature < 0.8  -> if (humidity < 0.5) Constants.plains else Constants.grassland
                     temperature <= 1.0 -> if (humidity < 0.7) Constants.desert else Constants.plains
                     else -> {
-                        println(temperature)
-                        Constants.lakes
+                        println("applyHumidityAndTemperature: Invalid temperature $temperature")
+                        Constants.grassland
                     }
                 }
                 tile.setTerrainTransients()
                 continue
             }
 
-            val matchingTerrain = ruleset.terrains.values.firstOrNull { terrain ->
-                terrain.uniqueObjects.any {
-                    it.placeholderText == "Occurs at temperature between [] and [] and humidity between [] and []"
-                            && it.params[0].toFloat() < temperature && temperature <= it.params[1].toFloat()
-                            && it.params[2].toFloat() < humidity && humidity <= it.params[3].toFloat()
-                }
+            val matchingTerrain = limitsMap.firstOrNull {
+                it.tempFrom < temperature && temperature <= it.tempTo
+                && it.humidFrom < humidity && humidity <= it.humidTo
             }
 
-            if (matchingTerrain != null) tile.baseTerrain = matchingTerrain.name
+            if (matchingTerrain != null) tile.baseTerrain = matchingTerrain.terrain.name
             else {
-                tile.baseTerrain = ruleset.terrains.keys.first()
-                println("Temperature: $temperature, humidity: $humidity")
+                tile.baseTerrain = ruleset.terrains.values.firstOrNull { it.type == TerrainType.Land }?.name ?: Constants.grassland
+                println("applyHumidityAndTemperature: No terrain found for temperature: $temperature, humidity: $humidity")
             }
             tile.setTerrainTransients()
         }
@@ -389,8 +440,8 @@ class MapGenerator(val ruleset: Ruleset) {
     private fun spawnIce(tileMap: TileMap) {
         tileMap.setTransients(ruleset)
         val temperatureSeed = randomness.RNG.nextInt().toDouble()
-        for (tile in tileMap.values) {
-            if (tile.baseTerrain !in Constants.sea || tile.terrainFeatures.any())
+        for (tile in tileMap.values.asSequence()) {
+            if (tile.baseTerrain !in Constants.sea || tile.terrainFeatures.isNotEmpty())
                 continue
 
             val randomTemperature = randomness.getPerlinNoise(tile, temperatureSeed, scale = tileMap.mapParameters.tilesPerBiomeArea.toDouble(), nOctaves = 1)
@@ -429,34 +480,45 @@ class MapGenerationRandomness{
     }
 
 
-    fun chooseSpreadOutLocations(number: Int, suitableTiles: List<TileInfo>, initialDistance: Int): ArrayList<TileInfo> {
-        for (distanceBetweenResources in initialDistance downTo 1) {
-            var availableTiles = suitableTiles.toList()
-            val chosenTiles = ArrayList<TileInfo>()
+    fun chooseSpreadOutLocations(number: Int, suitableTiles: List<TileInfo>, mapRadius: Int): ArrayList<TileInfo> {
+        // Determine sensible initial distance from number of desired placements and mapRadius
+        // empiric formula comes very close to eliminating retries for distance
+        val sparsityFactor = (suitableTiles.size.toDouble() / HexMath.getNumberOfTilesInHexagon(mapRadius)).pow(0.4)
+        val initialDistance = max(1, (mapRadius * 1.2 / number.toDouble().pow(0.4) * sparsityFactor + 0.5).toInt())
 
-            // If possible, we want to equalize the base terrains upon which
-            //  the resources are found, so we save how many have been
-            //  found for each base terrain and try to get one from the lowerst
-            val baseTerrainsToChosenTiles = HashMap<String, Int>()
-            for(tileInfo in availableTiles){
-                if(tileInfo.baseTerrain !in baseTerrainsToChosenTiles)
-                    baseTerrainsToChosenTiles[tileInfo.baseTerrain] = 0
-            }
+        // If possible, we want to equalize the base terrains upon which
+        //  the resources are found, so we save how many have been
+        //  found for each base terrain and try to get one from the lowest
+        val baseTerrainsToChosenTiles = HashMap<String, Int>()
+        for (tileInfo in suitableTiles){
+            if (tileInfo.baseTerrain !in baseTerrainsToChosenTiles)
+                baseTerrainsToChosenTiles[tileInfo.baseTerrain] = 0
+        }
+
+        for (distanceBetweenResources in initialDistance downTo 1) {
+            var availableTiles = suitableTiles
+            val chosenTiles = ArrayList<TileInfo>(number)
+
+            for (terrain in baseTerrainsToChosenTiles.keys)
+                baseTerrainsToChosenTiles[terrain] = 0
 
             for (i in 1..number) {
                 if (availableTiles.isEmpty()) break
                 val orderedKeys = baseTerrainsToChosenTiles.entries
                         .sortedBy { it.value }.map { it.key }
                 val firstKeyWithTilesLeft = orderedKeys
-                        .first { availableTiles.any { tile -> tile.baseTerrain== it} }
-                val chosenTile = availableTiles.filter { it.baseTerrain==firstKeyWithTilesLeft }.random(RNG)
+                        .first { availableTiles.any { tile -> tile.baseTerrain == it} }
+                val chosenTile = availableTiles.filter { it.baseTerrain == firstKeyWithTilesLeft }.random(RNG)
                 availableTiles = availableTiles.filter { it.aerialDistanceTo(chosenTile) > distanceBetweenResources }
                 chosenTiles.add(chosenTile)
-                baseTerrainsToChosenTiles[firstKeyWithTilesLeft] = baseTerrainsToChosenTiles[firstKeyWithTilesLeft]!!+1
+                baseTerrainsToChosenTiles[firstKeyWithTilesLeft] = baseTerrainsToChosenTiles[firstKeyWithTilesLeft]!! + 1
             }
             // Either we got them all, or we're not going to get anything better
             if (chosenTiles.size == number || distanceBetweenResources == 1) return chosenTiles
+//            if (MapGenerator.consoleOutput)
+//                println("chooseSpreadOutLocations: distance $distanceBetweenResources failed, trying one less")
         }
+        // unreachable due to last loop iteration always returning
         throw Exception("Couldn't choose suitable tiles for $number resources!")
     }
 }