… core/src/com/unciv/logic/map/mapgenerator/NaturalWonderGenerator.kt
… core/src/com/unciv/logic/map/mapgenerator/RiverGenerator.kt