Resolves #4786 - some Info over there, but then I started optimizing and added instrumentation to prove those optimizations.
The actual issue resolution is only not using terrains.first() upon a certain failure...

<details><summary>Measurements before optimization</summary>

MapGenerator run with parameters (Huge Hexagonal Pangaea, Seed 1628279571376, 6/2/0.7/0.59999996/0.4/0.05/0.1/0.0)
MapGenerator.MapLandmassGenerator took 20.67ms
Mountain-like generation for Mountain
Hill-like generation for Hill
MapGenerator.raiseMountainsAndHills took 90.75ms
Temperature: -0.3531763882369978, humidity: 0.18667701350657573
MapGenerator.applyHumidityAndTemperature took 81.57ms
MapGenerator.spawnLakesAndCoasts took 45.77ms
MapGenerator.spawnVegetation took 2.62ms
MapGenerator.spawnRareFeatures took 2.0ms
MapGenerator.spawnIce took 13.75ms
Natural Wonders for this game: [Krakatoa, El Dorado, Cerro de Potosi, Fountain of Youth, Mount Fuji]
MapGenerator.NaturalWonderGenerator took 22.29ms
MapGenerator.RiverGenerator took 49.55ms
MapGenerator.spreadResources took 516.11ms
MapGenerator.spreadAncientRuins took 29.38ms

MapGenerator run with parameters (Huge Hexagonal Pangaea, Seed 1628383638257, 3/1/0.62/0.44/0.15/0.05/0.1/-0.08)
MapGenerator.MapLandmassGenerator took 15.61ms
Mountain-like generation for Mountain
Hill-like generation for Hill
MapGenerator.raiseMountainsAndHills took 78.54ms
Temperature: -0.33561568074118525, humidity: 0.18225636915378696
MapGenerator.applyHumidityAndTemperature took 21.88ms
MapGenerator.spawnLakesAndCoasts took 40.89ms
MapGenerator.spawnVegetation took 2.93ms
MapGenerator.spawnRareFeatures took 3.19ms
MapGenerator.spawnIce took 11.74ms
Natural Wonders for this game: [Grand Mesa, Krakatoa, Great Barrier Reef, Barringer Crater, El Dorado]
MapGenerator.NaturalWonderGenerator took 24.95ms
MapGenerator.RiverGenerator took 24.63ms
MapGenerator.spreadResources took 586.80ms
MapGenerator.spreadAncientRuins took 35.81ms

MapGenerator run with parameters (Huge Hexagonal Pangaea, Seed 1628383671238, 6/2/0.78999996/0.65999997/0.4/0.05/0.1/-0.049999997)
MapGenerator.MapLandmassGenerator took 10.31ms
Mountain-like generation for Mountain
Hill-like generation for Hill
MapGenerator.raiseMountainsAndHills took 62.47ms
MapGenerator.applyHumidityAndTemperature took 17.12ms
MapGenerator.spawnLakesAndCoasts took 38.94ms
MapGenerator.spawnVegetation took 4.37ms
MapGenerator.spawnRareFeatures took 3.62ms
MapGenerator.spawnIce took 17.19ms
Natural Wonders for this game: [Barringer Crater, Old Faithful, Great Barrier Reef, Grand Mesa, El Dorado]
MapGenerator.NaturalWonderGenerator took 37.20ms
MapGenerator.RiverGenerator took 42.98ms
MapGenerator.spreadResources took 356.13ms
MapGenerator.spreadAncientRuins took 11.74ms
</details>
<details><summary>Measurements after optimization</summary>

MapGenerator run with parameters (Huge Hexagonal Pangaea, Seed 1628279571376, 6/2/0.7/0.59999996/0.4/0.05/0.1/0.0)
MapGenerator.MapLandmassGenerator took 16.62ms
Mountain-like generation for Mountain
Hill-like generation for Hill
MapGenerator.raiseMountainsAndHills took 52.56ms
applyHumidityAndTemperature: No terrain found for temperature: -0.3531763882369978, humidity: 0.18667701350657573
MapGenerator.applyHumidityAndTemperature took 10.84ms
MapGenerator.spawnLakesAndCoasts took 65.24ms
MapGenerator.spawnVegetation took 3.95ms
MapGenerator.spawnRareFeatures took 2.70ms
MapGenerator.spawnIce took 7.92ms
Natural Wonders for this game: [Cerro de Potosi, Old Faithful, El Dorado, Krakatoa, Barringer Crater]
MapGenerator.NaturalWonderGenerator took 31.69ms
MapGenerator.RiverGenerator took 15.90ms
MapGenerator.spreadResources took 229.71ms
MapGenerator.spreadAncientRuins took 15.63ms

MapGenerator run with parameters (Huge Hexagonal Pangaea, Seed 1628383638257, 3/1/0.62/0.44/0.15/0.05/0.1/-0.08)
MapGenerator.MapLandmassGenerator took 13.49ms
Mountain-like generation for Mountain
Hill-like generation for Hill
MapGenerator.raiseMountainsAndHills took 31.73ms
applyHumidityAndTemperature: No terrain found for temperature: -0.33561568074118525, humidity: 0.18225636915378696
MapGenerator.applyHumidityAndTemperature took 6.32ms
MapGenerator.spawnLakesAndCoasts took 25.57ms
MapGenerator.spawnVegetation took 2.63ms
MapGenerator.spawnRareFeatures took 2.31ms
MapGenerator.spawnIce took 5.79ms
Natural Wonders for this game: [Mount Fuji, Grand Mesa, Barringer Crater, Cerro de Potosi, Fountain of Youth]
MapGenerator.NaturalWonderGenerator took 33.25ms
MapGenerator.RiverGenerator took 48.11ms
MapGenerator.spreadResources took 232.72ms
MapGenerator.spreadAncientRuins took 19.12ms

MapGenerator run with parameters (Huge Hexagonal Pangaea, Seed 1628383671238, 6/2/0.79999995/0.65999997/0.4/0.05/0.1/-0.049999997)
MapGenerator.MapLandmassGenerator took 17.97ms
Mountain-like generation for Mountain
Hill-like generation for Hill
MapGenerator.raiseMountainsAndHills took 40.50ms
MapGenerator.applyHumidityAndTemperature took 4.25ms
MapGenerator.spawnLakesAndCoasts took 27.35ms
MapGenerator.spawnVegetation took 2.92ms
MapGenerator.spawnRareFeatures took 2.5ms
MapGenerator.spawnIce took 6.39ms
Natural Wonders for this game: [Rock of Gibraltar, Grand Mesa, Cerro de Potosi, Krakatoa, Mount Fuji]
MapGenerator.NaturalWonderGenerator took 20.89ms
MapGenerator.RiverGenerator took 34.95ms
MapGenerator.spreadResources took 192.74ms
MapGenerator.spreadAncientRuins took 8.42ms
</details>

Notes:
- The first test in the comparisons is ravignir's seed showing that one hill on a lake.
- Reproducibility of the timings is reduced a little by some generators not being deterministic by the given seed - vegetation and resources, for instance, will change when repeating a seed, coastlines lakes and mountains will not.
- applyHumidityAndTemperature is _way_ down, pity it's a small proportion anyway - those were the obvious optimizations.
- biggest gain is bringing the repetitions of the for loop in `chooseSpreadOutLocations` down, it often did start out with distance 40 and was finally succesful at distance 6... My formula succeeds but is a horrible monster. Better guesses welcome.
- I tried hard not to change resulting output at all. I passed up a few opportunities because proving nobody would notice any difference wouldn't have been trivial.

Draft only because I _expect_ someone to point out the BS.