@SomeTroglodyte 2021-08-08
Thanks Xander... There's minor things like
- ArrayList.any() is less efficient than ArrayList.isNotEmpty() because the former instantiates an iterator while the latter accesses materialized size, so it's about memory allocation count not nanoseconds
- arrayOf(const, const) inside loop - will the compiler be clever enough to compile this to static data or one allocation outside the loop or will it allocate each iteration? I don't trust it in this case to be clever enough - too many layers.
- Looking where 'materialization' of iterative code is necessary, where is a copy necessary, is shallow/deep needed (that one toList was a shallow copy and unnecessary 'cause input always already freshly copied, and restarting that loop didn't count)

and bigger ones like
- moving the parsing of the temp/humid ranges out of a loop if it's not dependent on a loop val
- reading that distance downto 1 loop as - each iteration starts from scratch and does not reuse results from the previous loop -> starting as low as possible is desirable
- doing some stats in Libre Calc with inputs + distance where the loop succeeded and then just _guessing_ a formula that matches (e.g. using pow(0.4) because I couldn't think quickly of an easy inversion to `n*(n+1)`)

So these changes _could_ change results _if_ my formula guesses the doable minimum distance lower than necessary. Since river head / resource / ruins placements are not controlled by the UI seed one cannot really test, or am I blind?

Also - where are better comments needed?