@SomeTroglodyte 2021-08-08
Strange. Mail says you commented on min/max, and can't find the comment or follow the mail's link... Maybe because max is right. But I stumble at exactly the same thing every time. Thing is, you _can_ get there with small maps, such as the map generation for main menu decoration. So that part is tested, too.