@SomeTroglodyte 2021-08-08
But part of that complicated formula should use HexMath. `getNumberOfTilesInHexagon` and `getEquivalentHexagonalRadius` are already there, and by splitting a `getHexagonalRadiusForArea` out of the latter the set is complete, and that should work to replace the `pow()`...