@SomeTroglodyte 2021-08-08
> Looks fine and mergable

->
> complicated formula should use HexMath

need to try, but was distracted (by there being too many and not enough const's in Constants). Plus, spurious leftovers.