@SomeTroglodyte 2021-08-08
Like it better now.

@ravignir : "No terrain found for temperature: -0.3856287273971997, humidity: 0.18030086524055133" still could be checked against the jsons
Last statistics:

|Size|Settings|Radius|suitable|number|Old guess|New guess|Actual|
|------|-----------|---------|-----------|--------|----------------|-----------|-----|
|Huge|Default|40|275|27|4|7|5|
|||40|1861|44|7|8|6|
|||40|2004|44|7|8|7|
|||40|1542|44|7|7|6|
|||40|1063|44|6|7|7|
|||40|1392|44|6|7|6|
|||40|2235|44|8|8|7|
|||40|2791|267|4|3|2|
|||40|2526|267|4|3|2|
|||40|2671|53|8|7|6|
||||||||
|Huge|Max|40|29|19|2|1|1|
|||40|76|123|1|1|1|
|||40|129|123|2|1|1|
|||40|6|123|1|1|1|
|||40|933|123|4|4|3|
|||40|2|123|1|1|1|
|||40|928|123|4|4|2|
|||40|1564|743|2|2|1|
|||40|966|743|2|1|1|
|||40|1486|29|8|9|8|
||||||||
|Huge|Min|40|329|31|4|7|5|
|||40|2750|13|14|16|15|
|||40|3001|13|14|16|16|
|||40|2307|13|13|16|14|
|||40|1118|13|10|14|15|
|||40|2087|13|12|15|15|
|||40|2952|13|14|16|15|
|||40|3491|78|7|6|5|
|||40|3426|78|7|6|5|
|||40|3148|62|8|7|6|