@SomeTroglodyte 2021-08-08
To all subscribers here - that routine I spent so much time on maximizes even spacing of resources / river sources / ruins, and I think the result can look quite unnatural. Try a huge map with all sliders to the left and look at the ruins.

Let's think about maybe a 'distribution uniformness' slider which has its 100% where the routine is working at now, and 0% switches it to run with no distance checks right away... And maybe an explicit resource cluster group size slider? Not hard to implement either. All the coal in the world clustered in groups of 1-8 (gauss-curved around the 4) could get interesting, no?